Move querybuilder related upsert code into query_builder module
