+++ diesel/src/query_builder/upsert/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod on_conflict_actions;
+pub(crate) mod on_conflict_clause;
+pub(crate) mod on_conflict_target;
